debugging-delete

+++ src/Controller/UserController.php
@@ -5,6 +5,7 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Entity\Customer;
 use Hateoas\HateoasBuilder;
+use OpenApi\Annotations as OA;
 use App\Repository\UserRepository;
 use App\Repository\CustomerRepository;
 use Doctrine\ORM\EntityManagerInterface;
@@ -15,11 +16,10 @@ use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\Serializer\Exception\NotEncodableValueException;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
-use OpenApi\Annotations as OA;
 
 
 
@@ -27,7 +27,7 @@ class UserController extends AbstractController
 {
     /**
      * @OA\Post(
-     *     path="/api/store_user",
+     *     path="/api/user",
      *   summary="Créer un nouvel utilisateur",
      *   @OA\RequestBody(
      *     required=true,
@@ -74,6 +74,10 @@ class UserController extends AbstractController
      *     response=401,
      *     description="JWT erreur de token"
      *   ),
+     *  @OA\Response(
+     *     response=403,
+     *     description="Accès interdit"
+     *   )
      * )
      * @Route("/api/user", name="store_user",methods={"POST"})
      */
@@ -115,26 +119,55 @@ class UserController extends AbstractController
 
      /**
       * @OA\Delete(
-     *     path="/api/delete_user/{id}",
+     *     path="/api/user/{id}/delete",
      *   summary="Supprimer un utilisateur par ID",
      *   @OA\Response(response=204, description="ok"),
      *   @OA\Response(response=401, description="Erreur du token jwt"),
-     *   @OA\Response(response=404, description="Aucun utilisateur avec cet id"),
+     *   @OA\Response(response=403, description="Vous n'êtes pas autorisé à accèder à cette ressource."),
+     *   @OA\Response(response=404, description="Cet utilisateur n'existe pas"),
+     *  @OA\Response(
+     *     response=403,
+     *     description="Accès interdit"
+     *   ),
      *   @OA\PathParameter(
      *     name="id",
      *     description="ID de l'utilisateur qui va être supprimé"
      *   )
      * )
-     * @IsGranted("ROLE_USER")
-     * @Route("/api/users/{id}/delete", name="delete_user",methods={"DELETE"})
+     * 
+     * @Route("/api/user/{id}/delete", name="delete_user")
      * 
      */
-    public function deleteUser(EntityManagerInterface $em,User $user)
+    public function deleteUser(int $id,EntityManagerInterface $em,CustomerRepository $customerRepository,UserRepository $userRepository)
     {
-        $em->remove($user);
-        $em->flush();
+        $customer = $customerRepository->find($this->getUser()->getId());
 
-        return new Response('ok');
+        $user = $userRepository->find($id);
+       
+        if ($user === null) {
+            return $this->json([
+                'status' => 404,
+                'message' => "Cet utilisateur n'existe pas"
+            ], 404);
+        }
+         if ($customer->getUsers() !== $this->getUser()) {
+            return $this->json([
+                'status' => 403,
+                'message' => "Vous n'êtes pas autorisé à accèder à cette ressource."
+            ], 403);
+        }
+
+        try {
+           $em->remove($user);
+           $em->flush();
+           
+        } catch (NotEncodableValueException $e) {
+            return $this->json([
+                'status' => $e->getCode(),
+                'message' => $e->getMessage()
+            ], 400);
+        }
+       
 
 
        
@@ -142,14 +175,14 @@ class UserController extends AbstractController
 
     /**
      * @OA\Get(
-     *     path="/api/users",
+     *     path="/users",
      *   summary="Liste des utilisateurs",
      *   @OA\Response(response=200, description="tous les utilisateurs"),
      *   @OA\Response(response=401, description="Erreur du token jwt"),
      *   @OA\Response(response=404, description="Aucun utilisateur trouvé")
      * )
-     * @IsGranted("ROLE_USER")
-     * @Route("/api/users", name="Users",methods={"GET"} )
+     * 
+     * @Route("/users", name="Users",methods={"GET"} )
      * 
      */
     public function listUser(UserRepository $userRepository): Response
@@ -171,11 +204,12 @@ class UserController extends AbstractController
      *     description="l'id de l'utilisateur que vous voulez recuperer"
      *   ),
      *   @OA\Response(response=200, description="Détail de l'utilisateur"),
+     *   @OA\Response(response=403, description="Vous n'êtes pas autorisé à accèder à cette ressource."),
      *   @OA\Response(response=401, description="Erreur du token JWT"),
-     *   @OA\Response(response=404, description="Aucun utilisateur trouvé avec cet ID")
+     *   @OA\Response(response=404, description="Cet utilisateur n'existe pas")
      * )
-     * @IsGranted("ROLE_USER")
-     * @Route("api/user/{id}", name="User",methods={"GET"})
+     * *
+     * @Route("/api/user/{id}", name="User",methods={"GET"})
      * 
      */
     public function findUser(UserRepository $userRepository,$id,SerializerInterface $serializer)
@@ -184,8 +218,30 @@ class UserController extends AbstractController
 
         $user = $userRepository->find($id);
  
-        $json = $hateoas->serialize($user, 'json', SerializationContext::create()->setGroups(array('customer:read')));
-        return new JsonResponse($json, 200, [], true);
+
+        if ($user === null) {
+            return $this->json([
+                'status' => 404,
+                'message' => "Cet utilisateur n'existe pas"
+            ], 404);
+        }
+         if ($user->getCustomer() !== $this->getUser()) {
+            return $this->json([
+                'status' => 403,
+                'message' => "Vous n'êtes pas autorisé à accèder à cette ressource."
+            ], 403);
+        }
+
+        try {
+            $json = $hateoas->serialize($user, 'json', SerializationContext::create()->setGroups(array('customer:read')));
+            return new JsonResponse($json, 200, [], true);
+        } catch (NotEncodableValueException $e) {
+            return $this->json([
+                'status' => $e->getCode(),
+                'message' => $e->getMessage()
+            ], 400);
+        }
+   
 
     }
 }
